Port arg parsing from shell script
