+++ g_cli/Cargo.toml
@@ -1,6 +1,8 @@
 [package]
-name = "g-lib"
-version = "0.1.0"
+name = "g"
+version = "0.4.0"
 authors = ["Ryan James Spencer <spencer.ryanjames@gmail.com>"]
 
 [dependencies]
+clap = "2.32.0"
+g_lib = { path = "../g_lib" }
