+++ g_lib/Cargo.toml
@@ -1,7 +1,6 @@
 [package]
-name = "g"
-version = "0.4.0"
+name = "g_lib"
+version = "0.1.0"
 authors = ["Ryan James Spencer <spencer.ryanjames@gmail.com>"]
 
 [dependencies]
-clap = "2.32.0"
